Flatpak: autodetect subtitle text encoding
Manually specifying text encoding is just wrong.
Especially for people who have no idea what text encoding they use (or what it is).
Lets try to be a little more user friendly and autodetect the encoding of each file.

The detection will be done inside GStreamer when first text buffer is received,
so another (but this time optional) patch is added, but who cares :-)