Fix not playing URIs
In case of local files the "filename" variable holds the string that is later assigned to the filename property (no difference which is used), but when playing URIs only filename property holds the string. Passing an undefined value to the label caused an error in this case and prevented playback of URIs.