clapper: playlist-demux: Improved "claps" typefinding
Suggesting "possible" type does not always work as intended, since lots
of text might be determined as "text/plain" due to higher probability.

We should give at least a "likely" typefind result instead of "possible",
but in order for that lets also check some more lines of playlist just to
be safe when announcing likeness.